requirements.txt
python-docx
pypdf2
PyQt6
qtpy
